Split out SVG files, write tests
