+++ src/lib/components/SVG/ErrorIcon.tsx
@@ -1,0 +1,16 @@
+import React from "react"
+
+export function ErrorIcon() {
+  return (
+    <svg
+      data-testid="error-icon"
+      className="error-icon"
+      focusable="false"
+      aria-hidden="true"
+      viewBox="0 0 24 24"
+    >
+      <path d="M15.73 3H8.27L3 8.27v7.46L8.27 21h7.46L21 15.73V8.27zM19 14.9 14.9 19H9.1L5 14.9V9.1L9.1 5h5.8L19 9.1z"></path>
+      <path d="M11 7h2v6h-2zm0 8h2v2h-2z"></path>
+    </svg>
+  )
+}
