+++ src/lib/components/SVG/FileSvg.tsx
@@ -1,0 +1,15 @@
+import React from "react"
+
+export function FileSvg() {
+  return (
+    <svg
+      data-testid="file-svg"
+      className="folder-upload-icon"
+      focusable="false"
+      aria-hidden="true"
+      viewBox="0 0 24 24"
+    >
+      <path d="M20 6h-8l-2-2H4c-1.1 0-1.99.9-1.99 2L2 18c0 1.1.9 2 2 2h16c1.1 0 2-.9 2-2V8c0-1.1-.9-2-2-2m0 12H4V8h16zM8 13.01l1.41 1.41L11 12.84V17h2v-4.16l1.59 1.59L16 13.01 12.01 9z"></path>
+    </svg>
+  )
+}
