+++ src/lib/components/SVG/InfoOutlinedIcon.tsx
@@ -1,0 +1,14 @@
+export function InfoOulinedIcon() {
+  return (
+    <svg
+      data-testid="info-outlined-icon"
+      className="info-oulined-icon"
+      focusable="false"
+      aria-hidden="true"
+      viewBox="0 0 24 24"
+    >
+      <path d="M11 7h2v2h-2zm0 4h2v6h-2z"></path>
+      <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2m0 18c-4.42 0-8-3.58-8-8s3.58-8 8-8 8 3.58 8 8-3.58 8-8 8"></path>
+    </svg>
+  )
+}
