+++ src/lib/components/SVG/index.tsx
@@ -1,100 +1,7 @@
-export function FileSvg() {
-  return (
-    <svg
-      data-testid="file-svg"
-      className="folder-upload-icon"
-      focusable="false"
-      aria-hidden="true"
-      viewBox="0 0 24 24"
-    >
-      <path d="M20 6h-8l-2-2H4c-1.1 0-1.99.9-1.99 2L2 18c0 1.1.9 2 2 2h16c1.1 0 2-.9 2-2V8c0-1.1-.9-2-2-2m0 12H4V8h16zM8 13.01l1.41 1.41L11 12.84V17h2v-4.16l1.59 1.59L16 13.01 12.01 9z"></path>
-    </svg>
-  )
-}
-
-export function CloseIcon() {
-  return (
-    <svg
-      data-testid="close-icon"
-      className="close-icon"
-      focusable="false"
-      aria-hidden="true"
-      viewBox="0 0 24 24"
-    >
-      <path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"></path>
-    </svg>
-  )
-}
-
-export function CheckCircleIcon() {
-  return (
-    <svg
-      data-testid="check-circle-icon"
-      className="check-circle-icon"
-      focusable="false"
-      aria-hidden="true"
-      viewBox="0 0 24 24"
-    >
-      <path d="M16.59 7.58 10 14.17l-3.59-3.58L5 12l5 5 8-8zM12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2m0 18c-4.42 0-8-3.58-8-8s3.58-8 8-8 8 3.58 8 8-3.58 8-8 8"></path>
-    </svg>
-  )
-}
-
-export function WarningTriangleOutlineIcon() {
-  return (
-    <svg
-      data-testid="warning-triangle-outline-icon"
-      className="warning-triangle-outline-icon"
-      focusable="false"
-      aria-hidden="true"
-      viewBox="0 0 24 24"
-    >
-      <path d="M12 5.99 19.53 19H4.47zM12 2 1 21h22z"></path>
-      <path d="M13 16h-2v2h2zm0-6h-2v5h2z"></path>
-    </svg>
-  )
-}
-
-export function ErrorIcon() {
-  return (
-    <svg
-      data-testid="error-icon"
-      className="error-icon"
-      focusable="false"
-      aria-hidden="true"
-      viewBox="0 0 24 24"
-    >
-      <path d="M15.73 3H8.27L3 8.27v7.46L8.27 21h7.46L21 15.73V8.27zM19 14.9 14.9 19H9.1L5 14.9V9.1L9.1 5h5.8L19 9.1z"></path>
-      <path d="M11 7h2v6h-2zm0 8h2v2h-2z"></path>
-    </svg>
-  )
-}
-
-export function InfoOulinedIcon() {
-  return (
-    <svg
-      data-testid="info-outlined-icon"
-      className="info-oulined-icon"
-      focusable="false"
-      aria-hidden="true"
-      viewBox="0 0 24 24"
-    >
-      <path d="M11 7h2v2h-2zm0 4h2v6h-2z"></path>
-      <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2m0 18c-4.42 0-8-3.58-8-8s3.58-8 8-8 8 3.58 8 8-3.58 8-8 8"></path>
-    </svg>
-  )
-}
-
-export function BellOutlinedIcon() {
-  return (
-    <svg
-      data-testid="bell-outlined-icon"
-      className="bell-outlined-icon"
-      focusable="false"
-      aria-hidden="true"
-      viewBox="0 0 24 24"
-    >
-      <path d="M12 22c1.1 0 2-.9 2-2h-4c0 1.1.9 2 2 2m6-6v-5c0-3.07-1.63-5.64-4.5-6.32V4c0-.83-.67-1.5-1.5-1.5s-1.5.67-1.5 1.5v.68C7.64 5.36 6 7.92 6 11v5l-2 2v1h16v-1zm-2 1H8v-6c0-2.48 1.51-4.5 4-4.5s4 2.02 4 4.5z"></path>
-    </svg>
-  )
-}
+export { BellOutlinedIcon } from "./BellOutlinedIcon"
+export { CheckCircleIcon } from "./CheckCircleIcon"
+export { CloseIcon } from "./CloseIcon"
+export { ErrorIcon } from "./ErrorIcon"
+export { FileSvg } from "./FileSvg"
+export { InfoOulinedIcon } from "./InfoOutlinedIcon"
+export { WarningTriangleOutlineIcon } from "./WarningTriangleOutlineIcon"
